Add generation result demo
